Refactor CMakeLists.txt and add header files for publisher and subscriber classes
