+++ samples/cpp/src/rde_cpp/src/rde_subscriber.cpp
@@ -1,26 +1,16 @@
-#include <memory>
+#include "rde_cpp/rde_subscriber.hpp"
 
-#include "rclcpp/rclcpp.hpp"
-#include "std_msgs/msg/string.hpp"
-
-class RdeSubscriber : public rclcpp::Node
+RdeSubscriber::RdeSubscriber()
+: Node("rde_subscriber")
 {
-public:
-  RdeSubscriber()
-  : Node("rde_subscriber")
-  {
-    subscription_ = this->create_subscription<std_msgs::msg::String>(
-      "rde", 10, std::bind(&RdeSubscriber::topic_callback, this, std::placeholders::_1));
-  }
+  subscription_ = this->create_subscription<std_msgs::msg::String>(
+    "rde", 10, std::bind(&RdeSubscriber::topic_callback, this, std::placeholders::_1));
+}
 
-private:
-  void topic_callback(const std_msgs::msg::String::SharedPtr msg) const
-  {
-    RCLCPP_INFO(this->get_logger(), "I heard: '%s'", msg->data.c_str());
-  }
-  
-  rclcpp::Subscription<std_msgs::msg::String>::SharedPtr subscription_;
-};
+void RdeSubscriber::topic_callback(const std_msgs::msg::String::SharedPtr msg) const
+{
+  RCLCPP_INFO(this->get_logger(), "I heard: '%s'", msg->data.c_str());
+}
 
 int main(int argc, char * argv[])
 {
